+++ .gitignore
@@ -1,10 +1,14 @@
 .DS_Store
 node_modules
+
 /dist
 
 
 # local env files
 .env.local
+.env.dev
+.env.prod
+
 .env.*.local
 
 # Log files
