--- components/footer.vue
@@ -1,10 +1,0 @@
-<template>
-  <v-footer> </v-footer>
-</template>
-<script>
-export default {
-  data() {
-    return {};
-  },
-};
-</script>
