+++ layouts/default.vue
@@ -6,20 +6,17 @@
         <nuxt />
       </v-container>
     </v-main>
-    <FooterSection />
   </v-app>
 </template>
 
 <script>
 // import FrontStyle from "@/assets/css/style.css"
 import { mapActions, mapGetters } from "vuex";
-import FooterSection from "@/components/footer.vue";
 import HeaderSection from "@/components/header.vue";
 
 export default {
   components: {
     // FrontStyle,
-    FooterSection,
     HeaderSection,
   },
   data() {
